move json io tests to crac-impl
